refactor(slack): split allowed route tests
